fix(#350): address Codex round-1 findings (global dedup keying, cancel wiring, file split)
- High: the shared SelectionCardFallback serves every chunk cell, so its
  dedup now keys on the document-GLOBAL range in both chunked sites; the
  fallback post re-derives the local range from the live tag at fire
  time. Regression test: same local range in two chunks posts both.
- Medium: selectionCardFallback.cancel() wired into both dismantleUIView
  implementations and both content-reload branches so no stale dedup
  state or pending post survives a content swap or teardown.
- Low: selection-notification routing + WI-12b mapping + #350 projection
  moved to TXTBridgeShared+SelectionMapping.swift (208 lines);
  TXTBridgeShared.swift back to 144 lines.

GH #1687
