+++ vreader/Views/Reader/TXTChunkedReaderBridge.swift
@@ -112,6 +112,9 @@ struct TXTChunkedReaderBridge: UIViewRepresentable {
         // any pending/in-flight locate bloom when the reader tears down, so a
         // queued work item doesn't fire against a detached table / cell.
         coordinator.cancelLandingBloom()
+        // Bug #350 (Codex round 1, Medium): drop any pending fallback card
+        // post + dedup state so nothing fires against a detached cell.
+        coordinator.selectionCardFallback.cancel()
     }
 
     // MARK: - Chunk Resolution (pure, testable)
@@ -351,6 +354,11 @@ struct TXTChunkedReaderBridge: UIViewRepresentable {
             context.coordinator.chunks = chunks
             context.coordinator.chunkStartOffsets = chunkStartOffsets
             context.coordinator.attrStringCache.removeAll()
+            // Bug #350 (Codex round 1, Medium): a chunk rebuild invalidates
+            // any pending fallback card post and its dedup state — global
+            // ranges from the old chunk layout must not suppress or fire
+            // against the new.
+            context.coordinator.selectionCardFallback.cancel()
             tableView.reloadData()
         }
 
@@ -1122,19 +1130,24 @@ struct TXTChunkedReaderBridge: UIViewRepresentable {
             delegate?.selectionDidChange(utf16Range: UTF16Range(startUTF16: globalStart, endUTF16: globalEnd))
             // Bug #350: the card no longer depends EXCLUSIVELY on UIKit
             // requesting an edit menu — a finalized non-empty selection
-            // posts via the debounced fallback (deduped against the
-            // editMenuForTextIn fast path; same chunk-offset translation).
-            selectionCardFallback.selectionChanged(range: nsRange) { [weak self, weak textView] armed in
-                guard let self, let textView,
-                      textView.selectedRange == armed else { return }
+            // posts via the debounced fallback. Codex round 1 (High): the
+            // fallback is SHARED across all chunk cells, so dedup must key
+            // on the document-GLOBAL range (the same local range in two
+            // different chunks is two distinct selections); the post
+            // re-derives the local range from the live tag at fire time.
+            let globalRange = NSRange(location: globalStart, length: nsRange.length)
+            selectionCardFallback.selectionChanged(range: globalRange) { [weak self, weak textView] armed in
+                guard let self, let textView else { return }
                 let idx = textView.tag
                 let offset = (idx >= 0 && idx < self.chunkStartOffsets.count)
                     ? self.chunkStartOffsets[idx]
                     : 0
+                let local = NSRange(location: armed.location - offset, length: armed.length)
+                guard textView.selectedRange == local else { return }
                 TXTBridgeShared.postSelectionNotification(
                     .readerSelectionPopoverRequested,
                     from: textView,
-                    range: armed,
+                    range: local,
                     chunkOffset: offset
                 )
             }
@@ -1166,9 +1179,10 @@ struct TXTChunkedReaderBridge: UIViewRepresentable {
             // `cellForRowAt`). Out-of-range tags fall back to
             // offset 0 — defensive but should never fire in
             // production.
-            // Bug #350: dedup against the debounced selection fallback.
-            if range.length > 0, selectionCardFallback.shouldMenuPathPost(range: range) {
-                selectionCardFallback.recordMenuPathPost(range: range)
+            // Bug #350: dedup against the debounced selection fallback,
+            // keyed on the document-GLOBAL range (the shared fallback
+            // serves every chunk cell — local ranges alias across chunks).
+            if range.length > 0 {
                 let chunkIndex = textView.tag
                 // Codex Gate 4 round 1 (Low): clamp on both ends.
                 // `textView.tag` is `Int` and could in principle be
@@ -1181,12 +1195,18 @@ struct TXTChunkedReaderBridge: UIViewRepresentable {
                 let chunkOffset = (chunkIndex >= 0 && chunkIndex < chunkStartOffsets.count)
                     ? chunkStartOffsets[chunkIndex]
                     : 0
-                TXTBridgeShared.postSelectionNotification(
-                    .readerSelectionPopoverRequested,
-                    from: textView,
-                    range: range,
-                    chunkOffset: chunkOffset
+                let globalRange = NSRange(
+                    location: chunkOffset + range.location, length: range.length
                 )
+                if selectionCardFallback.shouldMenuPathPost(range: globalRange) {
+                    selectionCardFallback.recordMenuPathPost(range: globalRange)
+                    TXTBridgeShared.postSelectionNotification(
+                        .readerSelectionPopoverRequested,
+                        from: textView,
+                        range: range,
+                        chunkOffset: chunkOffset
+                    )
+                }
             }
             return UIMenu(children: [])
         }
